"late"
so apparently you can submit an assignment in canvas after it's due
even if you've already submitted one. and so now I changed a couple
of things and forgot it was after the due date and it's marked as
late. I'm pretty sure I was even the first person in the class to
finish this assignment. I'm gonna vore something.